requirements.txt
torch
openai-whisper
pytest